major layout refactoring
