+++ src/pages/main.tsx
@@ -1,26 +1,17 @@
-import {
-  FunctionComponent,
-  PropsWithChildren,
-  useCallback,
-  useContext,
-  useEffect,
-  useRef,
-  useState,
-} from "react";
+import { useCallback, useContext, useEffect, useRef, useState } from "react";
 import Live from "./Live";
-import { getChannelInfo, getStorage, setStorage } from "../lib/chromeapi";
 import NoAuthPage from "./NoAuth";
-import isValidToken from "../lib/validateToken";
 import Loading from "../components/Loading";
 import Error from "./Error";
-import InvalidateToken from "../components/InvalidateToken";
-import { clientId, checkConnection, objToParams } from "../lib/lib";
+import { checkConnection } from "../lib/lib";
 import Layout from "../components/Layout";
 import LoadingContext from "../lib/LoadingContext";
+import TokenContext from "../lib/TokenContext";
+import isValidToken from "../lib/validateToken";
 
-const Main: FunctionComponent<PropsWithChildren<unknown>> = () => {
-  const { isLoading, setLoading } = useContext(LoadingContext);
-  const [tokenValid, setTokenValid] = useState(false);
+function Main() {
+  const { loading, setLoading } = useContext(LoadingContext);
+  const { tokenValid, setTokenValid } = useContext(TokenContext);
   const [showRUSure, setShowRUSure] = useState(false);
   const [connected, setConnected] = useState<boolean>(false);
   const dialogRef = useRef<HTMLDialogElement | null>(null);
@@ -31,7 +22,7 @@ const Main: FunctionComponent<PropsWithChildren<unknown>> = () => {
     setTokenValid(valid);
     setConnected(valid);
     setLoading(false);
-  }, [setLoading]);
+  }, [setLoading, setTokenValid]);
 
   useEffect(() => {
     (async () => {
@@ -48,7 +39,7 @@ const Main: FunctionComponent<PropsWithChildren<unknown>> = () => {
   if (showRUSure) window.scrollTo(0, 0);
 
   if (connected === false) {
-    if (isLoading === true) {
+    if (loading === true) {
       checkConnection().then(validateToken);
     } else if (tokenValid === false) {
       return <NoAuthPage />;
@@ -61,7 +52,7 @@ const Main: FunctionComponent<PropsWithChildren<unknown>> = () => {
     }
   }
 
-  if (isLoading) {
+  if (loading) {
     return <Loading />;
   }
 
@@ -74,34 +65,9 @@ const Main: FunctionComponent<PropsWithChildren<unknown>> = () => {
       }}
       show={showRUSure}
     >
-      <InvalidateToken
-        ref={dialogRef}
-        onChoice={async (invalidate) => {
-          if (invalidate) {
-            const token = (await getStorage("NowLive:Token")) || "";
-            try {
-              await fetch(
-                `https://id.twitch.tv/oauth2/revoke${objToParams({
-                  clientId,
-                  token,
-                })}`,
-                { method: "POST" },
-              );
-            } catch (err) {
-              console.error(err);
-            }
-            await setStorage("NowLive:Token", "");
-
-            setTokenValid(false);
-            return getChannelInfo();
-          }
-
-          return dialogRef.current?.close();
-        }}
-      />
       <Live />
     </Layout>
   );
-};
+}
 
 export default Main;
